refactor: move activations.py to folder layers/
File `activations.py` is moved to the folder `layers/`.
For retro-compatibility, it is still possible to call:
`deel.lip.activations.GroupSort2`
However, it is now preferred to call: `deel.lip.layers.GroupSort2`
